Refactor

+++ include/scaffold.hpp
@@ -6,19 +6,19 @@
 
 class Scaffold {
 private:
-    std::string name;
-    long long int start;
-    long long int stop;
+  std::string name;
+  int64_t start;
+  int64_t stop;
 
 public:
-    Scaffold();
-    Scaffold(std::string, long long int, long long int);
-    ~Scaffold() = default;
-    Scaffold& operator=(const Scaffold&) = default;
-    // Getters
-    std::string getName() const noexcept;
-    long long int getStart() const noexcept;
-    long long int getStop() const noexcept;
-    long long int getLength() const noexcept;
+  Scaffold();
+  Scaffold(std::string, int64_t, int64_t);
+  ~Scaffold() = default;
+  Scaffold &operator=(const Scaffold &) = default;
+  // Getters
+  std::string getName() const noexcept;
+  int64_t getStart() const noexcept;
+  int64_t getStop() const noexcept;
+  int64_t getLength() const noexcept;
 };
-#endif  // INCLUDE_SCAFFOLD_HPP_
+#endif // INCLUDE_SCAFFOLD_HPP_
